backend/requirements.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
